# golang/docs/swagger.yaml
basePath: /api
definitions:
  api.App:
    properties:
      name:
        example: 商城商品
        type: string
    type: object
...
  common.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  common.SuccResponse:
    properties:
      succ:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a API documentation for OPENAI.
  title: OPENAI
  version: "1.0"
paths:
  /v1/apps:
    get:
      description: Get all apps
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.App'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get all apps
      tags:
      - App
    post:
      consumes:
      - application/json
      description: Creates a new app with the provided data
      parameters:
      - description: App data
        in: body
        name: app
        required: true
        schema:
          $ref: '#/definitions/api.App'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.App'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Create a new app
      tags:
      - App
  /v1/apps/{app_id}:
    delete:
      consumes:
      - application/json
      description: Delete an app by app ID
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted app
          schema:
            $ref: '#/definitions/common.SuccResponse'
        "400":
          description: Invalid app ID
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Delete an app
      tags:
      - App
...
  /v1/apps/{id}:
    get:
      description: Get app by ID
      parameters:
      - description: App ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.App'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get app by ID
      tags:
      - App
    put:
      consumes:
      - application/json
      description: 根据提供的应用ID和更新的应用数据更新应用信息
      parameters:
      - description: 应用ID
        in: path
        name: app_id
        required: true
        type: string
      - description: 更新的应用数据
        in: body
        name: appData
        required: true
        schema:
          $ref: '#/definitions/api.App'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.App'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: 更新应用
      tags:
      - App
...
swagger: "2.0"